Decode on close
