+++ dubbo-rpc/dubbo-rpc-triple/src/main/java/org/apache/dubbo/rpc/protocol/tri/h12/grpc/GrpcHttp2ServerTransportListener.java
@@ -26,7 +26,6 @@ import org.apache.dubbo.remoting.http12.exception.DecodeException;
 import org.apache.dubbo.remoting.http12.exception.UnimplementedException;
 import org.apache.dubbo.remoting.http12.h2.H2StreamChannel;
 import org.apache.dubbo.remoting.http12.h2.Http2Header;
-import org.apache.dubbo.remoting.http12.h2.Http2InputMessage;
 import org.apache.dubbo.remoting.http12.h2.Http2TransportListener;
 import org.apache.dubbo.remoting.http12.message.MethodMetadata;
 import org.apache.dubbo.remoting.http12.message.StreamingDecoder;
@@ -122,19 +121,6 @@ public class GrpcHttp2ServerTransportListener extends GenericHttp2ServerTranspor
         }
         return invocation;
     }
-
-    @Override
-    protected void onError(Http2InputMessage message, Throwable throwable) {
-        try {
-            message.close();
-        } catch (Exception e) {
-            throwable.addSuppressed(e);
-        }
-        onError(throwable);
-    }
-
-    @Override
-    protected void onFinally(Http2InputMessage message) {}
 
     @Override
     protected GrpcStreamingDecoder getStreamingDecoder() {
